requirements.txt
Django==4.2.7
Pillow==10.0.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7
whitenoise==6.6.0
python-decouple==3.8
psycopg2-binary==2.9.7